SM-4697 did change in API to fetch team users
Did required changes for PR comments
