+++ app/src/main/kotlin/com/thundermaps/apilib/android/impl/SaferMeClientImpl.kt
@@ -15,7 +15,6 @@ import com.thundermaps.apilib.android.api.resources.SessionsResource
 import com.thundermaps.apilib.android.api.resources.ShapeResource
 import com.thundermaps.apilib.android.api.resources.TaskResource
 import com.thundermaps.apilib.android.api.resources.TeamResource
-import com.thundermaps.apilib.android.api.resources.TeamUsersResource
 import com.thundermaps.apilib.android.api.resources.TracedContactsResource
 import com.thundermaps.apilib.android.impl.resources.AttachmentResourceImpl
 import com.thundermaps.apilib.android.impl.resources.BrandResourceImpl
@@ -29,7 +28,6 @@ import com.thundermaps.apilib.android.impl.resources.SessionsImpl
 import com.thundermaps.apilib.android.impl.resources.ShapeResourceImpl
 import com.thundermaps.apilib.android.impl.resources.TasksImpl
 import com.thundermaps.apilib.android.impl.resources.TeamResourceImpl
-import com.thundermaps.apilib.android.impl.resources.TeamUsersResourceImpl
 import com.thundermaps.apilib.android.impl.resources.TracedContactsImpl
 import io.ktor.util.KtorExperimentalAPI
 import javax.inject.Inject
@@ -48,8 +46,7 @@ class SaferMeClientImpl @Inject constructor(
     private val shapeResourceImpl: ShapeResourceImpl,
     private val categoryResourceImpl: CategoryResourceImpl,
     private val formResourceImpl: FormResourceImpl,
-    private val attachmentResourceImpl: AttachmentResourceImpl,
-    private val teamUserResourceImpl: TeamUsersResourceImpl
+    private val attachmentResourceImpl: AttachmentResourceImpl
 ) : SaferMeClient {
     override val taskResource: TaskResource get() = TasksImpl(androidClient)
 
@@ -85,9 +82,6 @@ class SaferMeClientImpl @Inject constructor(
 
     override val attachmentResource: AttachmentResource
         get() = attachmentResourceImpl
-
-    override val teamUsersResource: TeamUsersResource
-        get() = teamUserResourceImpl
 
     override fun defaultParams(): RequestParameters = RequestParameters(
         customRequestHeaders = HashMap(),
