+++ app/src/main/kotlin/com/thundermaps/apilib/android/impl/resources/TeamResourceImpl.kt
@@ -1,5 +1,6 @@
 package com.thundermaps.apilib.android.impl.resources
 
+import androidx.annotation.VisibleForTesting
 import com.google.gson.Gson
 import com.thundermaps.apilib.android.api.com.thundermaps.isInternetAvailable
 import com.thundermaps.apilib.android.api.requests.RequestParameters
@@ -7,6 +8,7 @@ import com.thundermaps.apilib.android.api.resources.TeamResource
 import com.thundermaps.apilib.android.api.responses.models.Result
 import com.thundermaps.apilib.android.api.responses.models.ResultHandler
 import com.thundermaps.apilib.android.api.responses.models.Team
+import com.thundermaps.apilib.android.api.responses.models.TeamUser
 import com.thundermaps.apilib.android.impl.AndroidClient
 import io.ktor.client.call.HttpClientCall
 import io.ktor.client.call.call
@@ -40,14 +42,43 @@ class TeamResourceImpl @Inject constructor(
         val call = client.call(HttpRequestBuilder().takeFrom(requestBuilder).apply {
             method = HttpMethod.Get
             url(AndroidClient.baseUrlBuilder(parameters).apply {
-                encodedPath = "${encodedPath}$PATH?$FIELDS"
+                encodedPath = "${encodedPath}$TEAM_PATH?$TEAM_FIELDS"
             }.build())
         })
         return call
     }
 
+    private suspend fun getTeamUsersCall(
+        parameters: RequestParameters,
+        teamId: String
+    ): HttpClientCall {
+        val (client, requestBuilder) = androidClient.client(parameters)
+        val call = client.call(HttpRequestBuilder().takeFrom(requestBuilder).apply {
+            method = HttpMethod.Get
+            url(AndroidClient.baseUrlBuilder(parameters).apply {
+                encodedPath = "$encodedPath$TEAM_PATH/$teamId/$TEAM_USER_PATH?$TEAM_USER_FIELDS"
+            }.build())
+        })
+        return call
+    }
+
+    override suspend fun getTeamUsers(
+        parameters: RequestParameters,
+        teamId: String
+    ): Result<List<TeamUser>> {
+        if (!parameters.host.isInternetAvailable()) {
+            return resultHandler.handleException(UnknownHostException())
+        }
+        val call = getTeamUsersCall(parameters, teamId)
+        return resultHandler.processResult(call, gson)
+    }
+
     companion object {
-        const val PATH = "teams"
-        const val FIELDS = "fields=mapbox_username,mapbox_dataset_id,mapbox_access_token"
+        private const val TEAM_USER_PATH = "team_users"
+        @VisibleForTesting
+        const val TEAM_USER_FIELDS = "fields=first_name,last_name,email"
+
+        const val TEAM_PATH = "teams"
+        const val TEAM_FIELDS = "fields=mapbox_username,mapbox_dataset_id,mapbox_access_token"
     }
 }
