+++ app/src/test/kotlin/com/thundermaps/apilib/android/impl/SaferMeClientTest.kt
@@ -18,7 +18,6 @@ import com.thundermaps.apilib.android.impl.resources.SessionsImpl
 import com.thundermaps.apilib.android.impl.resources.ShapeResourceImpl
 import com.thundermaps.apilib.android.impl.resources.TasksImpl
 import com.thundermaps.apilib.android.impl.resources.TeamResourceImpl
-import com.thundermaps.apilib.android.impl.resources.TeamUsersResourceImpl
 import com.thundermaps.apilib.android.impl.resources.TracedContactsImpl
 import io.ktor.util.KtorExperimentalAPI
 import org.junit.After
@@ -44,7 +43,6 @@ internal class SaferMeClientTest {
     private val categoryResourceImpl = mock<CategoryResourceImpl>()
     private val formResourceImpl = mock<FormResourceImpl>()
     private val attachmentResourceImpl = mock<AttachmentResourceImpl>()
-    private val teamUserResourceImpl = mock<TeamUsersResourceImpl>()
     private lateinit var saferMeClient: SaferMeClient
 
     @Before
@@ -60,8 +58,7 @@ internal class SaferMeClientTest {
             shapeResourceImpl,
             categoryResourceImpl,
             formResourceImpl,
-            attachmentResourceImpl,
-            teamUserResourceImpl
+            attachmentResourceImpl
         )
     }
